Fix duplicated code and missing pbxproj wiring from the #7216 revert
The raw revert of #7216 reintroduced several unrelated code blocks
that had been extracted into sibling files purely to satisfy the
Swift file-length budget (not because they were todo-specific):
AppDelegate's authURLDebugSummary/feedWorkstreamTitle helpers,
TabManager's hashNotifications, and Workspace's surfaceKind/
hasUnreadNotification. Remove the reintroduced duplicates from the
main files and restore them in their extraction siblings.

Also wire KeyboardShortcutSettingsFileStore+SectionParsers.swift and
KeyboardShortcutSettings+SystemWideHotkeyConflicts.swift into
project.pbxproj. Both files were already committed on this branch but
never added to the Sources build phase, so Xcode silently skipped
compiling them: KeyboardShortcutSettingsFileStore.swift called
parseFileEditorSection/parseFileExplorerSection but their definitions
in the unwired extraction file were invisible to the build, breaking
activation-session-benchmark and app-host unit tests in CI.
